Config comments
Unsure on a couple items. Also unsure if a moderator is able to revive a snitch, or if it's an admin as listed in the config. Also unsure if there is a namelayer permission associated to juke culling. Also unsure if you need to type /jainfo to stop the culling.